updated Input,Update,playableCharacter
### Since Last Commit
---
---
#### Additions
- N/A
---
#### Changes
- change Vector2f to Vector2i
---
#### Removed
- N/A

+++ vsSolutions/shadowOperative_001/shadowOperative_001/LevelManager.h
@@ -17,7 +17,7 @@
 class LevelManager {
 private:
 	sf::Vector2i m_LevelSize;
-	sf::Vector2f m_StartPosition;
+	sf::Vector2i m_StartPosition;
 	sf::Vector2i m_EndPosition;
 	std::list<sf::Vector2i> m_EnemyPosition;
 	std::list<sf::Vector2i> m_ItemPosition;
